Rename file
casper_contract_interface.rs -> generator.rs
