+++ macros/src/contract/generator.rs
