+++ macros/src/contract/mod.rs
@@ -1,9 +1,9 @@
 mod caller;
 mod casper_contract;
-mod casper_contract_interface;
 mod contract;
 mod contract_test;
+mod generator;
 mod parser;
 mod utils;
-pub use casper_contract_interface::generate_code;
+pub use generator::generate_code;
 pub use parser::CasperContractItem;
